Декомпозирует редьюсер appData на редьюсеры: offerListData, offerData, reviewData
