+++ project/src/store/review-data/selectors.ts
@@ -1,0 +1,11 @@
+import { NameSpace } from '../../const';
+import { State } from '../../types/state';
+import { Reviews } from '../../types/reviews';
+
+export const getReviews = (state: State): Reviews => state[NameSpace.Review].reviews;
+
+export const getReviewSendStatus = (state: State): boolean => state[NameSpace.Review].isReviewSending;
+
+export const getReviewSendSuccessStatus = (state: State): boolean => state[NameSpace.Review].isReviewSendSuccess;
+
+export const getErrorMessage = (state: State): string | null => state[NameSpace.Review].errorMessage;
