+++ project/src/types/state.ts
@@ -3,27 +3,23 @@ import { store } from '../store';
 import { Offer, Offers } from './offers';
 import { Reviews } from './reviews';
 
-export type UserProcess = {
-  authorizationStatus: AuthorizationStatus,
-  userData: {
-    email: string | null
-  },
-  isLogoutProcessing: boolean,
-  isLogoutError: boolean,
-  isLoginError: boolean
-};
-
 export type AppProcess = {
   activeCity: City
 }
 
-export type AppData = {
+export type OfferListData = {
   offersList: Offers,
   isOffersListLoading: boolean,
+}
+
+export type OfferData = {
   offer: Offer | null,
-  nearOffers: Offers,
-  reviews: Reviews,
   isOfferDataLoading: boolean,
+  nearOffers: Offers,
+}
+
+export type ReviewData = {
+  reviews: Reviews,
   isReviewSending: boolean,
   isReviewSendSuccess: boolean,
   errorMessage: string | null,
@@ -33,6 +29,16 @@ export type FavoriteData = {
   favoriteList: Offers,
   isFavoriteListLoading: boolean
 }
+
+export type UserProcess = {
+  authorizationStatus: AuthorizationStatus,
+  userData: {
+    email: string | null
+  },
+  isLogoutProcessing: boolean,
+  isLogoutError: boolean,
+  isLoginError: boolean
+};
 
 export type State = ReturnType<typeof store.getState>;
